requirements.txt
fastapi==0.70.0
pydantic==1.8.2
python-multipart==0.0.5
uvicorn==0.15.0
tensorflow-cpu==2.6.0
tensorflow-hub==0.12.0
tensorflow-text==2.6.0
opencv-python==4.5.4.58